+++ parser/src/Parse/Type.hs
@@ -20,17 +20,17 @@ tvar elmVersion =
 
 tuple :: ElmVersion -> IParser AST.Type
 tuple elmVersion =
-  addLocation $
+  addLocation $ checkMultiline $
   do  types <- parens'' (withEol $ expr elmVersion)
       case types of
         Left comments ->
-            return $ AST.UnitType comments
+            return $ \_ -> AST.UnitType comments
         Right [] ->
-            return $ AST.UnitType []
+            return $ \_ -> AST.UnitType []
         Right [AST.Commented [] (AST.WithEol t Nothing) []] ->
-            return $ A.drop t
+            return $ \_ -> A.drop t
         Right [AST.Commented pre (AST.WithEol t eol) post] ->
-            return $ AST.TypeParens (AST.Commented pre t (maybeToList (fmap AST.LineComment eol) ++ post))
+            return $ \_ -> AST.TypeParens (AST.Commented pre t (maybeToList (fmap AST.LineComment eol) ++ post))
         Right types' ->
             return $ AST.TupleType types'