+++ src/AST/MapNamespace.hs
@@ -40,8 +40,8 @@ instance MapNamespace Type' where
             TypeParens nested ->
                 TypeParens (mapNamespace f nested)
 
-            TupleType entries ->
-                TupleType (mapNamespace f entries)
+            TupleType entries forceMultiline ->
+                TupleType (mapNamespace f entries) forceMultiline
 
             RecordType base fields trailingComments forceMultiline ->
                 RecordType base (mapNamespace f fields) trailingComments forceMultiline